Split IR2 data types up into separate files
