+++ lib/Besra/Types/IR2/Number.hs
@@ -1,0 +1,8 @@
+
+module Besra.Types.IR2.Number ( Number(..) ) where
+
+import Protolude
+
+
+newtype Number = Number Int
+  deriving (Eq, Show)
